requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
configparser==3.5.0
Flask==1.0.2
Flask-Redis==0.3.0
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
libgravatar==0.2.3
MarkupSafe==1.1.0
more-itertools==5.0.0
pluggy==0.8.1
py==1.7.0
PyMySQL==0.9.3
pytest==4.1.1
python-http-client==3.1.0
redis==3.0.1
requests==2.21.0
sendgrid==5.6.0
six==1.12.0
soupsieve==1.7.2
SQLAlchemy==1.2.16
urllib3==1.24.1
Werkzeug==0.14.1
